backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyPDF2
docx2txt
langgraph
google-generativeai
pymongo
python-dotenv
